Test refactoring to create a "general" hybrid agents test section and move common code to a central location
